1 and 2 work
So calling dot operator on main method works, but not on methods in the same class or a different class.
Also we need to implement this now.  The main changes:
how new works
How M-name works and how var-in-scope works (had to handl ethe case for where dot was passed in)
How dot works for M-value
Look at changes here and in my 4.3 update for a full list of changes in code